+++ scriptabit/habitica_service.py
@@ -69,7 +69,8 @@ class HabiticaService(object):
     def is_server_up(self):
         """Check that the Habitica API is reachable and up
 
-        Returns: bool: `True` if the server is reachable, otherwise `False`.
+        Returns:
+            bool: `True` if the server is reachable, otherwise `False`.
         """
         response = self.__get('status')
         if response.status_code == requests.codes.ok:
@@ -79,7 +80,8 @@ class HabiticaService(object):
     def get_user(self):
         """Gets the authenticated user data.
 
-        Returns: dictionary: The user data.
+        Returns:
+            dict: The user data.
         """
         response = self.__get('user')
         response.raise_for_status()
@@ -88,7 +90,8 @@ class HabiticaService(object):
     def get_stats(self):
         """Gets the authenticated user stats.
 
-        Returns: dictionary: The stats.
+        Returns:
+            dict: The stats.
         """
         return self.get_user()['stats']
 
@@ -99,7 +102,8 @@ class HabiticaService(object):
             task_type (HabiticaTaskTypes): The type of task to get.
                 Default is all tasks.
 
-        Returns: dictionary: The tasks.
+        Returns:
+            dict: The tasks.
         """
         params = {'type': task_type.value} if task_type else {}
         response = self.__get('tasks/user', params)
@@ -115,7 +119,8 @@ class HabiticaService(object):
                 Default is to create a new todo. Only used if the task['type']
                 is empty or not present.
 
-        Returns: dict: The new task as returned from the server.
+        Returns:
+            dict: The new task as returned from the server.
         """
 
         if not task.get('type', None):
@@ -141,7 +146,8 @@ class HabiticaService(object):
             _id (str): The task ID.
             alias (str): The task alias.
 
-        Returns: dict: The task, or None if the task is not found.
+        Returns:
+            dict: The task, or None if the task is not found.
 
         Raises:
             ValueError
@@ -176,7 +182,8 @@ class HabiticaService(object):
                 task is required. Can be overriden by an existing task['type']
                 value.
 
-        Returns: dict: The new task as returned from the server.
+        Returns:
+            dict: The new task as returned from the server.
 
         Raises:
             ValueError
@@ -229,7 +236,8 @@ class HabiticaService(object):
         Args:
             hp (float): The new HP value.
 
-        Returns: float: The new HP value, extracted from the JSON response data.
+        Returns:
+            float: The new HP value, extracted from the JSON response data.
         """
         if hp > 50:
             raise ArgumentOutOfRangeError("hp > 50")
@@ -246,7 +254,8 @@ class HabiticaService(object):
         Args:
             mp (float): The new MP value.
 
-        Returns: float: The new MP value, extracted from the JSON response data.
+        Returns:
+            float: The new MP value, extracted from the JSON response data.
         """
         max_mp = self.get_user()['stats']['mp']
         if mp > max_mp:
@@ -264,7 +273,8 @@ class HabiticaService(object):
         Args:
             exp (float): The new XP value.
 
-        Returns: float: The new XP value, extracted from the JSON response data.
+        Returns:
+            float: The new XP value, extracted from the JSON response data.
         """
         if exp < 0:
             raise ArgumentOutOfRangeError("exp < 0")
@@ -279,7 +289,8 @@ class HabiticaService(object):
         Args:
             gp (float): The new gold value.
 
-        Returns: float: The new gold value, extracted from the response data.
+        Returns:
+            float: The new gold value, extracted from the response data.
         """
         if gp < 0:
             raise ArgumentOutOfRangeError("gp < 0")