+++ scriptabit/plugins/trello/trello.py
@@ -156,7 +156,7 @@ If empty, then cards are only marked done when archived.''')
         Returns: bool: True if further updates are required; False if the plugin
         is finished and the application should shut down.
         """
-        # logging.getLogger(__name__).info('')
+        logging.getLogger(__name__).debug('Getting Trello data')
 
         # retrieve the boards to sync
         boards = self.__tc.list_boards(board_filter='open')
@@ -188,13 +188,15 @@ If empty, then cards are only marked done when archived.''')
 
         # Create the services
         source_service = TrelloTaskService(self.__tc, sync_lists, done_lists)
-        assert self.__habitica_task_service
 
         # synchronise
+        logging.getLogger(__name__).debug('Starting Trello to Habitica sync')
         sync = TaskSync(source_service, self.__habitica_task_service, task_map)
         sync.synchronise(
             clean_orphans=True,
-            sync_completed_new_tasks=True)
+            sync_completed_new_tasks=False)
+
+        logging.getLogger(__name__).debug('Finished Trello to Habitica sync')
 
         # Persist the updated task map
         task_map.persist(self.__task_map_file)