+++ CHANGELOG.rst
@@ -110,4 +110,5 @@ Changelog
 -----------------------------------------
 
 * Fixed task sync bug with missing destination tasks.
+* Implemented Habitica sync task